chore: rename emulator hook files (#118)
