+++ gbajs3/src/components/controls/virtual-controls.spec.tsx
@@ -8,7 +8,7 @@ import { VirtualControls } from './virtual-controls.tsx';
 import { renderWithContext } from '../../../test/render-with-context.tsx';
 import { GbaDarkTheme } from '../../context/theme/theme.tsx';
 import * as contextHooks from '../../hooks/context.tsx';
-import * as quickReloadHooks from '../../hooks/emulator/quick-reload.tsx';
+import * as quickReloadHooks from '../../hooks/emulator/use-quick-reload.tsx';
 import { UploadSaveToServerModal } from '../modals/upload-save-to-server.tsx';
 
 import type { GBAEmulator } from '../../emulator/mgba/mgba-emulator.tsx';
