+++ gbajs3/src/components/modals/load-local-rom.spec.tsx
@@ -5,7 +5,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { LoadLocalRomModal } from './load-local-rom.tsx';
 import { renderWithContext } from '../../../test/render-with-context.tsx';
 import * as contextHooks from '../../hooks/context.tsx';
-import * as runGameHooks from '../../hooks/emulator/run-game.tsx';
+import * as runGameHooks from '../../hooks/emulator/use-run-game.tsx';
 import { productTourLocalStorageKey } from '../product-tour/consts.tsx';
 
 import type { GBAEmulator } from '../../emulator/mgba/mgba-emulator.tsx';
