+++ gbajs3/src/components/modals/upload-rom.spec.tsx
@@ -10,7 +10,7 @@ import { UploadRomModal } from './upload-rom.tsx';
 import { testRomLocation } from '../../../test/mocks/handlers.ts';
 import { renderWithContext } from '../../../test/render-with-context.tsx';
 import * as contextHooks from '../../hooks/context.tsx';
-import * as runGameHooks from '../../hooks/emulator/run-game.tsx';
+import * as runGameHooks from '../../hooks/emulator/use-run-game.tsx';
 import { productTourLocalStorageKey } from '../product-tour/consts.tsx';
 
 import type { GBAEmulator } from '../../emulator/mgba/mgba-emulator.tsx';
