+++ gbajs3/src/components/modals/upload-rom.tsx
@@ -17,7 +17,7 @@ import { ModalBody } from './modal-body.tsx';
 import { ModalFooter } from './modal-footer.tsx';
 import { ModalHeader } from './modal-header.tsx';
 import { useEmulatorContext, useModalContext } from '../../hooks/context.tsx';
-import { useRunGame } from '../../hooks/emulator/run-game.tsx';
+import { useRunGame } from '../../hooks/emulator/use-run-game.tsx';
 import { useLoadExternalRom } from '../../hooks/use-load-external-rom.tsx';
 import {
   EmbeddedProductTour,
