+++ gbajs3/src/hooks/emulator/use-background-emulator.tsx
