+++ gbajs3/src/hooks/emulator/use-quit-game.tsx
